fix(university): give university pages the same RepoPulse header as the main app
Extract AppHeader from ResultsShell into a shared component so both university
pages (/university and /university/[slug]) render the full dark nav header with
logo, Scoring Methodology link, GitHub link, theme toggle, and user badge —
instead of the plain DemoBanner that made them look like a separate app.
